restore setting dns record values in the libdns.Record.Value fields
This keeps setting Priority/Preference/Weight on libdns.Record based on the dns.RR.
But when going from libdns.Record to dns.RR, this just keeps the fields from
the Value, ignoring the Priority/Preference/Weight fields on the libdns.Record.

This also keeps the fix of turning an empty relative name in a libdns.Record
into "@" before parsing it as a dns record. Without that change, setting
records at the zone apex would fail (which will be detected properly as a
FORMERR with PR #10).

For issue #11